v1.2.1
Changed code so it no longer requires the roundabout way to resize - Simple Fix involving changing an "open()" to an "Image.open()"
Otherwise, nothing else done due to inability to combat #48
This is because to update an element with a gif involves ".UpdateAnimation" (or similar) yet only works if it is in the main loop so it can continue updating.
This isn't possible at the current moment, I will likely do some more research into it but at the current moment, #48 shall remain unfixed and marked as wontfix.